Created weekText file from AppShell.jsx and created tests for weekText
