+++ www/scripts/build.js
@@ -37,7 +37,7 @@ function handleErrors(stats) {
     // eslint-disable-next-line max-len
     printErrors(
       'Failed to compile. When process.env.CI = true, warnings are treated as failures. Most CI servers set this automatically.',
-      stats.compilation.warnings,
+      stats.compilation.warnings
     );
     process.exit(1);
   }
